requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
redis==5.0.2
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
asttokens==2.4.1
base58==0.2.5
colorama==0.4.6
executing==2.0.1
icecream==2.1.3
pygments==2.17.2
six==1.16.0
pickleDB~=0.9.2
fastecdsa==2.3.2
shortuuid==1.0.12
pymongo==4.6.2
python-dotenv==1.0.1